Fix two bugs
The exit handler unit tests were unit testing one of the assembly
functions that has privileged instructions in it which caused a seg
fault. This code does not need unit tested so I simply removed the test.

The second bug was with the memory manager. The memory manager was
compiling the malloc functions in Windows which it should not be as
these are only needed by the cross comnpiled version of the VMM, so I
added the VMM macro to prevent that code from being compiled while its
is being tested.

Signed-off-by: Rian Quinn <“rianquinn@gmail.com”>